requirements.txt
appnope==0.1.0
asgiref==3.2.7
backcall==0.1.0
colored-traceback==0.3.0
decorator==4.4.2
Django==3.0
django-tempus-dominus==5.1.2.13
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.9
